+++ src/pyiem/webutil.py
@@ -106,11 +106,15 @@ XSS_SENTINEL = "XSS"
 MEMCACHED_HIT = "_mhit"
 
 
-def _conv2list(mixed) -> list:
-    """Convert to a list."""
+def _conv2list(mixed: str | list) -> list:
+    """Ensure we have a list of strings without csv delimination."""
     if isinstance(mixed, list):
-        return mixed
-    return mixed.split(",")
+        # pydantic my come here with a string that is still delimited
+        result = []
+        for item in mixed:
+            result.extend(x.strip() for x in item.split(","))
+        return result
+    return [x.strip() for x in mixed.split(",")]
 
 
 def _ensure_all_strings(mixed) -> list: