+++ tests/test_webutil.py
@@ -497,6 +497,30 @@ def test_iemapp_times_notime():
     assert list(application(env, sr))[0].decode("ascii").find("Hello") > -1
 
 
+def test_listorcsvtype_provided_list_with_csv():
+    """Test that we flatten this situation."""
+
+    class MyModel(CGIModel):
+        """Test."""
+
+        wfo: ListOrCSVType = Field(None)
+
+    res = MyModel(wfo=["BGM,DMX"])
+    assert res.wfo == ["BGM", "DMX"]
+
+
+def test_listorcsvtype_provided_list_with_csv_and_other():
+    """Test that we flatten this situation."""
+
+    class MyModel(CGIModel):
+        """Test."""
+
+        wfo: ListOrCSVType = Field(None)
+
+    res = MyModel(wfo=["BGM,DMX", "DVN"])
+    assert res.wfo == ["BGM", "DMX", "DVN"]
+
+
 def test_iemapp_bracket_variable():
     """Test that a bracked variable is handled within pydantic schema."""