+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this library are documented in this file.
 ### New Features
 
 - Account for SPC PTS one-off with reversed and closed polygon.
+- Account for FastAPI usage of `webutil.ListOrCSVType`.
 - Add `ip_throttle_secs` to `webutil.iemapp` to deal with IEM pain.
 - Gracefully handle `XTEUS` product without a value set.
 - Improve `iemapp` to better capture actual HTTP status_code and document